+++ src/command_modules/azure-cli-container/azure/cli/command_modules/container/_params.py
@@ -15,7 +15,7 @@ from ._validators import validate_volume_mount_path, validate_secrets, validate_
 
 # pylint: disable=line-too-long
 
-IP_ADDRESS_TYPES = ['Public', 'Private']
+IP_ADDRESS_TYPES = ['Public']
 
 
 def _environment_variables_type(value):
@@ -69,7 +69,7 @@ def load_arguments(self, _):
         c.argument('environment_variables', nargs='+', options_list=['--environment-variables', '-e'], type=_environment_variables_type, help='A list of environment variable for the container. Space-separated values in \'key=value\' format.')
         c.argument('secure_environment_variables', nargs='+', type=_secure_environment_variables_type, help='A list of secure environment variable for the container. Space-separated values in \'key=value\' format.')
         c.argument('secrets', secrets_type)
-        c.argument('network_profile', help='The network profile id for VNET')
+        c.argument('network_profile', help='The network profile id')
         c.argument('secrets_mount_path', validator=validate_volume_mount_path, help="The path within the container where the secrets volume should be mounted. Must not contain colon ':'.")
         c.argument('file', options_list=['--file', '-f'], help="The path to the input file.")